♻️ Refactor: 이미지 경로 변경 및 타이틀 변경
