+++ src/constants/categoryName.js
@@ -9,16 +9,16 @@ export const COOKING = '요리';
 export const ANIMALS = '동물';
 export const ETC = '기타';
 
-import movie from '../assets/movie.png';
-import travel from '../assets/travel.png';
-import art from '../assets/art.png';
-import music from '../assets/music.png';
-import diy from '../assets/diy.png';
-import exercise from '../assets/exercise.png';
-import game from '../assets/game.png';
-import cooking from '../assets/cooking.png';
-import animal from '../assets/animal.png';
-import etc from '../assets/etc.png';
+import movie from '../../public/movie.png';
+import travel from '../../public/travel.png';
+import art from '../../public/art.png';
+import music from '../../public/music.png';
+import diy from '../../public/diy.png';
+import exercise from '../../public/exercise.png';
+import game from '../../public/game.png';
+import cooking from '../../public/cooking.png';
+import animal from '../../public/animal.png';
+import etc from '../../public/etc.png';
 
 // 카테고리 이름과 이미지를 객체 배열로 만듦
 export const categoryArr = [
