+++ src/pages/SignUp.jsx
@@ -125,7 +125,7 @@ const Signup = () => {
           <div className="user-image">
             <img
               className="logo-img"
-              src="/src/assets/test-logo.png"
+              src="/team_logo.png"
               alt="site_logo"
               onClick={() => navigate('/')}
             />
